Task 13

+++ 0x02-functions_nested_loops/101-natural.c
@@ -7,7 +7,7 @@
 */
 int main(void)
 {
-	int i, sum, sum1, sum2;
+	unsigned long int i, sum, sum1, sum2;
 
 	sum = 0;
 	sum1 = 0;
